Add "Go to page" form.
This commit adds the ability to type a page identifier into a small form at the bottom
of boltd and it will immediately go to that page. This is not safe for use while the
database is in use as the page could be changed and could point to an overflow page.